seperate recent and saved renderlist

+++ python/BlackPepper/ui/controller.py
@@ -29,11 +29,10 @@ class PepperWindow(QMainWindow):
         PepperWindow 실행 시 self.login_ui가 우선 실행된다.
         """
         super().__init__()
-        self.preset_json_path = ''
-        self.recent_menu = None
-        self.saved_menu = None
         self.pepper = Houpub()
         self.login_log = Auto_log()
+        self.recent_menu = None
+        self.saved_menu = None
         self.projects_selection = None
         self.templates_selection = None
         self.shots_selection = None
@@ -143,8 +142,8 @@ class PepperWindow(QMainWindow):
         log_sfw = self.login_window.hipbox.currentText()[1:]
         log_value = self.login_log.load_setting()
         log_dict = self.login_log.user_dict
-        if os.path.exists(log_path) and(not log_dict['auto'] or log_id != log_value['user_id']
-                                        or log_pw != log_value['user_pw'] or log_sfw != log_value['user_ext']):
+        if os.path.exists(log_path) and (not log_dict['auto'] or log_id != log_value['user_id']
+                                         or log_pw != log_value['user_pw'] or log_sfw != log_value['user_ext']):
             self.login_log.user_id = log_id
             self.login_log.user_pw = log_pw
             self.login_log.user_ext = log_sfw
@@ -378,9 +377,6 @@ class PepperWindow(QMainWindow):
         return
 
     def renew_render_list(self):
-        # self.render_model.pepperlist.clear()
-        # for render in self.pepper.precomp_list:
-        #     self.render_model.pepperlist.append(render)
         self.shots_selection.clear()
         self.render_model.layoutChanged.emit()
         self.renderlists_selection.clear()
@@ -436,16 +432,11 @@ class PepperWindow(QMainWindow):
         # create 'File' menu
         self.main_filemenu = self.main_menu_bar.addMenu('&File')
         # 동적으로 메뉴를 채워주는 부분
-        self.main_filemenu.aboutToShow.connect(self.set_main_open_recent)
-        self.main_filemenu.aboutToShow.connect(self.set_main_open_saved)
+        self.main_filemenu.aboutToShow.connect(self.set_main_menubar)
         # create Help menu
         main_helpmenu = self.main_menu_bar.addMenu('&Help')
         # 'File' menu add 'Open Recent preset' & 'Exit'
-        self.set_main_open_recent()
-        # self.set_main_open_saved()
-
-        # self.main_filemenu.addAction()
-        # main_helpmenu = self.main_menu_bar.addMenu('Help')
+        self.set_main_menubar()
         # set kitsu
         kisu_action = QAction('Kitsu', self.main_window)
         kisu_action.setShortcut('F1')
@@ -478,49 +469,34 @@ class PepperWindow(QMainWindow):
         logout_action.triggered.connect(self.user_logout)
         self.main_user.addAction(logout_action)
 
-    def set_main_open_recent(self):
+    def set_main_menubar(self):
         """메인창의 file 메뉴 'Open Recent Presets' 의 sub 메뉴들을 만들어주는 함수이다.
 
             메뉴바 'Menu' 에 'Recent Presets는'메뉴에 path json file 을 불러오고 최신 5개의 json render_model.pepperlist
             5번 인덱스가 제일 최신 json dict 정보이다.
         """
         self.main_filemenu.clear()
-
-        if not os.path.exists(self.preset_json_path):
-            self.create_json()
-        pass
+        self.create_json()
 
         self.recent_menu = QMenu('Open recent renderlists', self.main_window)
+        self.saved_menu = QMenu('Open saved renderlists', self.main_window)
 
         with open(self.preset_json_path, 'r') as f:
             self.render_list_data = json.load(f)
 
-        # if not os.path.exists(self.preset_json_path):
-        #     self.presave_preset_json()
-        #     return
-        # else:
         for json_files in self.render_list_data['recent']:
             for file_path in json_files:
                 file_action = QAction(os.path.basename(file_path), self)
                 file_action.triggered.connect(lambda _, path=file_path: self.handle_file(path))
                 self.recent_menu.addAction(file_action)
 
-        self.main_filemenu.addMenu(self.recent_menu)
-
-        if not os.path.exists(self.preset_json_path):
-            self.create_json()
-        pass
-
-        self.saved_menu = QMenu('Open saved renderlists', self.main_window)
-
-        with open(self.preset_json_path, 'r') as f:
-            self.saved_list_data = json.load(f)
-
-        for json_files in self.saved_list_data['saved']:
+        for json_files in self.render_list_data['saved']:
             for file_path in json_files:
                 file_action = QAction(os.path.basename(file_path), self)
                 file_action.triggered.connect(lambda _, path=file_path: self.handle_file(path))
-                self.recent_menu.addAction(file_action)
+                self.saved_menu.addAction(file_action)
+
+        self.main_filemenu.addMenu(self.recent_menu)
         self.main_filemenu.addMenu(self.saved_menu)
 
         # add 'Exit'
@@ -530,27 +506,12 @@ class PepperWindow(QMainWindow):
         exit_action.triggered.connect(QApplication.instance().quit)
         self.main_filemenu.addAction(exit_action)
 
-    def set_main_open_saved(self):
-        """메인창의 메뉴바에 Recent precet 메뉴를 만들어주 함수이다.
-
-            메뉴바 'Menu' 에 'Recent Presets는'메뉴에 path json file 을 불러오고 최신 5개의 json render_model.pepperlist
-            5번 인덱스가 제일 최신 json dict 정보이다.
-        """
-        if not os.path.exists(self.preset_json_path):
-            self.create_json()
-        pass
-
-        self.saved_menu = QMenu('Open saved renderlists', self.main_window)
-
-        with open(self.preset_json_path, 'r') as f:
-            self.saved_list_data = json.load(f)
-
-        for json_files in self.saved_list_data['saved']:
+    def append_renderlist_to_menubar(self, data_type):
+        for json_files in self.render_list_data[data_type]:
             for file_path in json_files:
                 file_action = QAction(os.path.basename(file_path), self)
                 file_action.triggered.connect(lambda _, path=file_path: self.handle_file(path))
-                self.recent_menu.addAction(file_action)
-        self.main_filemenu.addMenu(self.saved_menu)
+                self.saved_menu.addAction(file_action)
 
     def handle_file(self, file_path):
         """메인창의 file 메뉴 'Open Recent Presets' 의 원하는 Preset 을 선택시 실행되는 함수이다.
@@ -567,10 +528,8 @@ class PepperWindow(QMainWindow):
         """
         if len(self.render_model.pepperlist) == 0:
             return
-        if not os.path.exists(self.preset_json_path):
-            self.create_json()
-        with open(self.preset_json_path, 'r') as f:
-            self.render_list_data = json.load(f)
+        self.create_json()
+        self.open_json()
         recent_data = self.render_list_data.get('recent', [])
         now = datetime.now()
         # 최대 인덱스 5까지 새로운 value가 추가되도록 수정
@@ -581,9 +540,7 @@ class PepperWindow(QMainWindow):
         })
         # 'recent' key 값의 value로 저장
         self.render_list_data['recent'] = recent_data
-        data_to_save = self.render_list_data
-        with open(self.preset_json_path, "w") as f:
-            json.dump(data_to_save, f, ensure_ascii=False)
+        self.save_json(self.render_list_data)
 
     def save_user_renderlists(self):
         """save preset json path 의 json 을 불러오고 recent key 값에 정보를 저장하는 함수이다.
@@ -593,10 +550,8 @@ class PepperWindow(QMainWindow):
         """
         if len(self.render_model.pepperlist) == 0:
             return
-        if not os.path.exists(self.preset_json_path):
-            self.create_json()
-        with open(self.preset_json_path, 'r') as f:
-            self.render_list_data = json.load(f)
+        self.create_json()
+        self.open_json()
         saved_data = self.render_list_data.get('saved', [])
         now = datetime.now()
         # 최대 인덱스 5까지 새로운 value가 추가되도록 수정
@@ -607,21 +562,28 @@ class PepperWindow(QMainWindow):
         })
         # 'recent' key 값의 value로 저장
         self.render_list_data['saved'] = saved_data
-        data_to_save = self.render_list_data
+        self.save_json(self.render_list_data)
+
+    def open_json(self):
+        with open(self.preset_json_path, 'r') as f:
+            self.render_list_data = json.load(f)
+
+    def save_json(self, data):
         with open(self.preset_json_path, "w") as f:
-            json.dump(data_to_save, f, ensure_ascii=False)
+            json.dump(data, f, ensure_ascii=False)
 
     def create_json(self):
         """preset이 저장되어있는 json파일이 없으면 json 파일을 만들어주는 함수이다.
         """
-        self.render_list_data = {
-            "recent": [],
-            "saved": []
-        }
-        data_to_save = self.render_list_data
+        if not os.path.exists(self.preset_json_path):
+            self.render_list_data = {
+                "recent": [],
+                "saved": []
+            }
+            data_to_save = self.render_list_data
 
-        with open(self.preset_json_path, "w") as f:
-            json.dump(data_to_save, f, ensure_ascii=False)
+            with open(self.preset_json_path, "w") as f:
+                json.dump(data_to_save, f, ensure_ascii=False)
 
     def render_execute(self):
         houp = HouPepper()
